test/regression/reftest/vendor/blink/fast/dom/interface-object-proto.html

<html>
<head>
<script src="../../resources/wpt-test-pre.js"></script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>

<script>
description("Tests the prototype chain of interface objects");

shouldBe('Object.getPrototypeOf(HTMLBodyElement)', 'HTMLElement');
shouldBe('Object.getPrototypeOf(HTMLElement)', 'Element');
shouldBe('Object.getPrototypeOf(Element)', 'Node');
shouldBe('Object.getPrototypeOf(Node)', 'EventTarget');
shouldBe('Object.getPrototypeOf(EventTarget)', 'Function.prototype');

</script>
<script src="../../resources/js-test-post.js"></script>
<script src="../../resources/wpt-test-post.js"></script>
</body>
</html>
